task/python: add save/restore last tool/pocket prepped

+++ configs/sim/pysubs/sqltooltable.py
@@ -13,6 +13,8 @@ class SqlToolTable(object):
         print "SQL tt init"
         self.inifile = inifile
         self.random_toolchanger = random_toolchanger
+        self.persist =  int(self.inifile.find("TOOL", "SAVE_TOOLSTATE") or 0)
+
         self.connectstring = self.inifile.find("TOOL", "ODBC_CONNECT")
         self.conn = pyodbc.connect(self.connectstring)
         self.cursor = self.conn.cursor()
@@ -50,7 +52,6 @@ class SqlToolTable(object):
                 if t.toolno != -1: print str(t)
 
     def save(self, tooltable, comments,fms):
-        print "SQL save",
         try:
             self.cursor.execute("delete from tools;")
             start = 0 if self.random_toolchanger else 1
@@ -61,19 +62,18 @@ class SqlToolTable(object):
                                         (t.toolno, p, t.diameter, t.backangle,t.frontangle,t.orientation,
                                          'xxx',t.offset.x,t.offset.y,t.offset.z,t.offset.a,t.offset.b,
                                          t.offset.c,t.offset.u,t.offset.v,t.offset.w))
-
             self.cursor.execute("commit;")
-            print "done"
 
         except pyodbc.Error, msg:
+            print "saving tooltable failed:"
             traceback.print_exc(file=sys.stdout)
-            print "saving tooltable to %s failed" % (self.filename)
 
 
     def save_state(self,e):
+        if not self.persist:
+            return
         print "SQL save_state",
         try:
-
             self.cursor.execute("delete from state;")
             self.cursor.execute("insert into state values(?,?);",e.io.tool.toolInSpindle,e.io.tool.pocketPrepped)
             self.cursor.execute("commit;")
@@ -85,10 +85,10 @@ class SqlToolTable(object):
         pass
 
     def restore_state(self,e):
-        print "SQL save_state",
+        if not self.persist:
+            return
         try:
-            self.cursor.execute("select tool_in_spindle,pocket_prepped from state;")
-            row = self.cursor.fetchone()
+            row = self.cursor.execute("select tool_in_spindle,pocket_prepped from state;").fetchone()
             e.io.tool.pocketPrepped = row.pocket_prepped
             e.io.tool.toolInSpindle = row.tool_in_spindle
             print "restored tool=%d pocket=%d" % (row.tool_in_spindle,row.pocket_prepped)
@@ -96,5 +96,4 @@ class SqlToolTable(object):
         except pyodbc.Error, msg:
             print "restore_state() failed:"
             traceback.print_exc(file=sys.stdout)
-        pass
         pass
